requirements.txt
discord.py >= 2.0.0
certifi == 2022.9.24
google >= 3.0.0
gspread >= 5.6.0
python-dotenv >= 0.21.0